enable -Wconversion and fix a warning

+++ examples/select.cpp
@@ -107,7 +107,7 @@ int main()
 				.from(p.join(x).on(p.feature == x.feature))
 				.where(true)))
 	{
-		int id = row.id;
+		int64_t id = row.id;
 		std::string name = row.name;
 	}
 #endif
